Rename headerTags module js to match pattern name
