+++ styleguide/source/assets/js/index.js
@@ -9,6 +9,7 @@ import eventFilters               from "./modules/eventFilters";
 import eventListingInteractive    from "./modules/eventListingInteractive";
 import footnote                   from "./modules/footnote.js";
 import formValidation             from "./modules/formValidation.js";
+import headerTags                 from "./modules/headerTags.js";
 import hideAlert                  from "./modules/hideAlert.js";
 import keywordSearch              from "./modules/keywordSearch.js";
 import locationFilters            from "./modules/locationFilters.js";
@@ -18,7 +19,6 @@ import mobileNav                  from "./modules/mobileNav.js";
 import orgSelector                from "./modules/orgSelector.js";
 import pagination                 from "./modules/pagination.js";
 import pikaday                    from "./modules/pikaday.js";
-import relationshipIndicators     from "./modules/relationshipIndicators.js";
 import responsiveVideo            from "./modules/responsiveVideo.js";
 import resultsHeading             from "./modules/resultsHeading.js";
 import richText                   from "./modules/richText.js";
